+++ plugins/tracker-resources/src/components/sprints/SprintDatePresenter.svelte
@@ -1,0 +1,48 @@
+<!--
+// Copyright © 2022 Hardcore Engineering Inc.
+// 
+// Licensed under the Eclipse Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License. You may
+// obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// 
+// See the License for the specific language governing permissions and
+// limitations under the License.
+-->
+<script lang="ts">
+  import { Timestamp } from '@anticrm/core'
+
+  import { getClient } from '@anticrm/presentation'
+  import { Sprint } from '@anticrm/tracker'
+  import { DatePresenter } from '@anticrm/ui'
+
+  export let value: Sprint
+  export let field = 'targetDate'
+  export let kind: 'transparent' | 'primary' | 'link' | 'list' = 'primary'
+
+  const client = getClient()
+
+  $: dateMs = (value as any)[field] as number
+
+  const handleDateChanged = async (event: CustomEvent<Timestamp>) => {
+    const newDate = event.detail
+
+    if (newDate === undefined || dateMs === newDate) {
+      return
+    }
+
+    await client.update(value, { [field]: newDate })
+  }
+</script>
+
+<DatePresenter
+  value={dateMs}
+  editable={true}
+  shouldShowLabel={true}
+  icon={'normal'}
+  {kind}
+  on:change={handleDateChanged}
+/>