+++ plugins/tracker-resources/src/components/sprints/Sprints.svelte
@@ -1,0 +1,59 @@
+<!--
+// Copyright © 2022 Hardcore Engineering Inc.
+// 
+// Licensed under the Eclipse Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License. You may
+// obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// 
+// See the License for the specific language governing permissions and
+// limitations under the License.
+-->
+<script lang="ts">
+  import { Ref } from '@anticrm/core'
+  import { IntlString } from '@anticrm/platform'
+  import { createQuery } from '@anticrm/presentation'
+  import { Sprint, Team } from '@anticrm/tracker'
+  import { closePopup, closeTooltip, location } from '@anticrm/ui'
+  import { onDestroy } from 'svelte'
+  import tracker from '../../plugin'
+  import { SprintViewMode } from '../../utils'
+  import EditSprint from './EditSprint.svelte'
+  import SprintBrowser from './SprintBrowser.svelte'
+
+  export let currentSpace: Ref<Team>
+  export let label: IntlString = tracker.string.Sprints
+  export let search: string = ''
+  export let mode: SprintViewMode = 'all'
+
+  let sprintId: Ref<Sprint> | undefined
+  let sprint: Sprint | undefined
+
+  onDestroy(
+    location.subscribe(async (loc) => {
+      closeTooltip()
+      closePopup()
+
+      sprintId = loc.path[5] as Ref<Sprint>
+    })
+  )
+
+  const sprintQuery = createQuery()
+  $: if (sprintId !== undefined) {
+    sprintQuery.query(tracker.class.Sprint, { _id: sprintId }, (result) => {
+      sprint = result.shift()
+    })
+  } else {
+    sprintQuery.unsubscribe()
+    sprint = undefined
+  }
+</script>
+
+{#if sprint}
+  <EditSprint {sprint} />
+{:else}
+  <SprintBrowser {label} query={{ space: currentSpace }} {search} {mode} />
+{/if}